+++ src/shared/ui/PriceInfoContainer/PriceBox/PriceTextBox/index.tsx
